Modify the interface by introducing BigDecimal type of variable..

+++ src/main/java/com/stream/bigdecimal/SandStorage.java
@@ -1,5 +1,7 @@
 package com.stream.bigdecimal;
 
+import java.math.BigDecimal;
+
 public interface SandStorage {
-    long getSandBeansQuantity();
+    BigDecimal getSandBeansQuantity();
 }
